% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{br_micro_codes}
\alias{br_micro_codes}
\title{List of all micro-regions in Brazil}
\usage{
br_micro_codes
}
\description{
List of all micro-regions in Brazil and associated mesoregion, state, and region codes and names.
}
