% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{br_muni_codes}
\alias{br_muni_codes}
\title{List of all municipalities in Brazil}
\usage{
br_muni_codes
}
\description{
List of all municipalities in Brazil and associated municipality, microregion, mesoregion, state, and region codes and names.
}
