% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{muni_codes}
\alias{muni_codes}
\title{List of all municipalities in Brazil}
\usage{
muni_codes
}
\description{
List of all municipalities in Brazil and associated municipality, microregion, mesoregion, and region codes and names.
}
